+++ Sources/Resources/MeshAnimated.h
@@ -16,7 +16,7 @@ class MeshAnimated : public Mesh
 		MeshAnimated(const std::string& meshName,					const bool& isAnimable,                
 					 const std::vector<glm::vec3>& verticesArray,	const std::vector<glm::vec3>& normalesArray, 
 					 const std::vector<glm::vec3>& colorArray,		const std::vector<glm::ivec3>& bonesArray,   
-					 const std::vector<glm::vec3>& weightsArray,	const std::vector<unsigned int>& facesArray);
+					 const std::vector<glm::vec3>& weightsArray,	const std::vector<unsigned short>& facesArray);
         ~MeshAnimated();
 		//
